# src/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "FurWaz API Documentation",
    "description": "Documentation for the FurWaz API",
    "version": "1.0.0",
    "contact": {
      "name": "FurWaz",
      "url": "https://furwaz.fr",
      "email": "contact@furwaz.fr"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "authorization"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/apps/": {
      "get": {
        "tags": [
          "Apps"
        ],
        "description": "Get all apps",
        "operationId": "getApps",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Apps"
        ],
        "description": "Create a new app",
        "operationId": "createApp",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/apps/me": {
      "get": {
        "tags": [
          "Apps"
        ],
        "description": "Get the apps that the user owns",
        "operationId": "getOwnApps",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/apps/{id}": {
      "get": {
        "tags": [
          "Apps"
        ],
        "description": "Get an app by its ID",
        "operationId": "getAppById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Apps"
        ],
        "description": "Update an app by its ID",
        "operationId": "updateAppById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Apps"
        ],
        "description": "Delete an app by its ID",
        "operationId": "deleteAppById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/portal/{portal}/app": {
      "get": {
        "tags": [
          "Portal"
        ],
        "description": "Get the portal app",
        "operationId": "getPortalApp",
        "parameters": [
          {
            "name": "portal",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/store/buy/": {
      "post": {
        "tags": [
          "Product"
        ],
        "description": "Request a token for user product addition",
        "operationId": "requestBuyToken",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "productId": {
                  "example": "any"
                },
                "userId": {
                  "example": "any"
                },
                "quantity": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/store/buy/{token}/infos": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "Get product infos",
        "operationId": "getBuyInfos",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/store/buy/{token}/status": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "Get product status",
        "operationId": "getBuyStatus",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/store/buy/{token}": {
      "post": {
        "tags": [
          "Product"
        ],
        "description": "Add product to cart",
        "operationId": "addProductToCart",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/store/cart/": {
      "get": {
        "tags": [
          "Cart"
        ],
        "description": "Get own user cart",
        "operationId": "getUserProducts",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Cart"
        ],
        "description": "Empty own user cart",
        "operationId": "deleteUserProducts",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/store/cart/product/{productId}": {
      "get": {
        "tags": [
          "Cart"
        ],
        "description": "Get cart product",
        "operationId": "getUserProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Cart"
        ],
        "description": "Update cart product",
        "operationId": "updateUserProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Cart"
        ],
        "description": "Delete cart product",
        "operationId": "deleteUserProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    }
  }
}